+++ code/frontend/rental/src/views/Listing.vue
@@ -68,6 +68,11 @@
 
 <script>
 import axios from 'axios';
+import { signIn } from '@aws-amplify/auth';
+import { signOut } from '@aws-amplify/auth';
+import "@aws-amplify/ui-vue/styles.css";
+import { fetchAuthSession } from "aws-amplify/auth";
+import { Amplify } from "aws-amplify";
 export default {
   name: 'ListingPage',
   data() {
@@ -106,6 +111,36 @@ export default {
     };
   },
   methods: {
+    async login() {
+      try {
+        const user = await signIn({
+          username: 'ding874946686@gmail.com',
+          password: 'D13607161848@jc',
+        });
+        console.log('Login successful:', user);
+        const currentConfig = Amplify.getConfig();
+        Amplify.configure({
+          ...currentConfig,
+          API: {
+            REST: {
+              RentalNinja: {
+                endpoint: "https://api.rentalninja.link",
+                region: "us-east-2",
+              },
+            },
+          },
+        });
+
+        const { accessToken, idToken } = (await fetchAuthSession()).tokens ?? {};
+        localStorage.setItem('idToken', idToken);
+        localStorage.setItem('ac', accessToken);
+        await signOut();
+        
+      } catch (error) {
+        console.error('Login error:', error);
+        this.errorMessage = 'Login failed: ' + error.message; // 设置错误反馈
+      }
+    },
     async get_post() {
       try {
         const idToken = localStorage.getItem('idToken');
@@ -180,6 +215,7 @@ export default {
   mounted() {
     this.applyFilters();
     this.get_post();
+    this.login();
   },
 };
 </script>
